Fixed a number of testing bugs

+++ src/main/java/org/dasein/cloud/test/compute/ComputeResources.java
@@ -94,13 +94,15 @@ public class ComputeResources {
         this.provider = provider;
     }
 
-    public void report() {
+    public int report() {
         boolean header = false;
+        int count = 0;
 
         testMachineImages.remove(DaseinTestManager.STATELESS);
         if( !testMachineImages.isEmpty() ) {
             logger.info("Provisioned Compute Resources:");
             header = true;
+            testMachineImages.size();
             DaseinTestManager.out(logger, null, "---> Machine Images", testMachineImages.size() + " " + testMachineImages);
         }
         testSnapshots.remove(DaseinTestManager.STATELESS);
@@ -109,6 +111,7 @@ public class ComputeResources {
                 logger.info("Provisioned Compute Resources:");
                 header = true;
             }
+            count += testSnapshots.size();
             DaseinTestManager.out(logger, null, "---> Snapshots", testSnapshots.size() + " " + testSnapshots);
         }
         testVMs.remove(DaseinTestManager.STATELESS);
@@ -117,6 +120,7 @@ public class ComputeResources {
                 logger.info("Provisioned Compute Resources:");
                 header = true;
             }
+            count += testVMs.size();
             DaseinTestManager.out(logger, null, "---> Virtual Machines", testVMs.size() + " " + testVMs);
         }
         testVolumes.remove(DaseinTestManager.STATELESS);
@@ -124,12 +128,15 @@ public class ComputeResources {
             if( !header ) {
                 logger.info("Provisioned Compute Resources:");
             }
+            count+= testVolumes.size();
             DaseinTestManager.out(logger, null, "---> Volumes", testVolumes.size() + " " + testVolumes);
         }
+        return count;
     }
 
-    public void close() {
+    public int close() {
         ComputeServices computeServices = provider.getComputeServices();
+        int count = 0;
 
         if( computeServices != null ) {
             VirtualMachineSupport vmSupport = computeServices.getVirtualMachineSupport();
@@ -138,10 +145,18 @@ public class ComputeResources {
                 for( Map.Entry<String,String> entry : testVMs.entrySet() ) {
                     if( !entry.getKey().equals(DaseinTestManager.STATELESS) ) {
                         try {
-                            vmSupport.terminate(entry.getValue());
+                            VirtualMachine vm = vmSupport.getVirtualMachine(entry.getValue());
+
+                            if( vm != null ) {
+                                vmSupport.terminate(entry.getValue());
+                                count++;
+                            }
+                            else {
+                                count++;
+                            }
                         }
-                        catch( Throwable ignore ) {
-                            // ignore
+                        catch( Throwable t ) {
+                            logger.warn("Failed to de-provision test VM " + entry.getValue() + ": " + t.getMessage());
                         }
                     }
                 }
@@ -153,10 +168,18 @@ public class ComputeResources {
                 for( Map.Entry<String,String> entry : testMachineImages.entrySet() ) {
                     if( !entry.getKey().equals(DaseinTestManager.STATELESS) ) {
                         try {
-                            imageSupport.remove(entry.getValue());
+                            MachineImage img = imageSupport.getImage(entry.getValue());
+
+                            if( img != null ) {
+                                imageSupport.remove(entry.getValue());
+                                count++;
+                            }
+                            else {
+                                count++;
+                            }
                         }
-                        catch( Throwable ignore ) {
-                            // ignore
+                        catch( Throwable t ) {
+                            logger.warn("Failed to de-provision test image " + entry.getValue() + ": " + t.getMessage());
                         }
                     }
                 }
@@ -168,10 +191,18 @@ public class ComputeResources {
                 for( Map.Entry<String,String> entry : testSnapshots.entrySet() ) {
                     if( !entry.getKey().equals(DaseinTestManager.STATELESS) ) {
                         try {
-                            snapshotSupport.remove(entry.getValue());
+                            Snapshot snapshot = snapshotSupport.getSnapshot(entry.getValue());
+
+                            if( snapshot != null ) {
+                                snapshotSupport.remove(entry.getValue());
+                                count++;
+                            }
+                            else {
+                                count++;
+                            }
                         }
                         catch( Throwable t ) {
-                            logger.warn("Failed to deprovision snapshot " + entry.getValue() + " post-test: " + t.getMessage());
+                            logger.warn("Failed to de-provision test snapshot " + entry.getValue() + " post-test: " + t.getMessage());
                         }
                     }
                 }
@@ -182,10 +213,14 @@ public class ComputeResources {
                 for( Map.Entry<String,String> entry : testVolumes.entrySet() ) {
                     if( !entry.getKey().equals(DaseinTestManager.STATELESS) ) {
                         try {
-                            volumeSupport.detach(entry.getValue(), true);
+                            Volume volume = volumeSupport.getVolume(entry.getValue());
+
+                            if( volume != null ) {
+                                volumeSupport.detach(entry.getValue(), true);
+                            }
                         }
                         catch( Throwable ignore ) {
-                            // ignore
+                            // IGNORE
                         }
                     }
                 }
@@ -194,16 +229,25 @@ public class ComputeResources {
                 for( Map.Entry<String,String> entry : testVolumes.entrySet() ) {
                     if( !entry.getKey().equals(DaseinTestManager.STATELESS) ) {
                         try {
-                            volumeSupport.remove(entry.getValue());
+                            Volume volume = volumeSupport.getVolume(entry.getValue());
+
+                            if( volume != null ) {
+                                volumeSupport.remove(entry.getValue());
+                                count++;
+                            }
+                            else {
+                                count++;
+                            }
                         }
-                        catch( Throwable ignore ) {
-                            // ignore
+                        catch( Throwable t ) {
+                            logger.warn("Failed to de-provision test volume " + entry.getValue() + ": " + t.getMessage());
                         }
                     }
                 }
             }
         }
         provider.close();
+        return count;
     }
 
     private @Nullable String findStatelessSnapshot() {
@@ -330,7 +374,7 @@ public class ComputeResources {
     public @Nullable String getTestSnapshotId(@Nonnull String label, boolean provisionIfNull) {
         if( label.equals(DaseinTestManager.STATELESS) ) {
             for( Map.Entry<String,String> entry : testSnapshots.entrySet() ) {
-                if( !entry.getKey().equals(DaseinTestManager.REMOVED) ) {
+                if( !entry.getKey().startsWith(DaseinTestManager.REMOVED) ) {
                     String id = entry.getValue();
 
                     if( id != null ) {
@@ -693,9 +737,47 @@ public class ComputeResources {
         SnapshotCreateOptions options;
 
         if( volumeId == null ) {
-            volumeId = getTestVolumeId(label, true, null, null);
+            volumeId = getTestVolumeId(DaseinTestManager.STATEFUL + (System.currentTimeMillis()%1000), true, null, null);
             if( volumeId == null ) {
                 throw new CloudException("No volume from which to create a snapshot");
+            }
+        }
+        @SuppressWarnings("ConstantConditions") VolumeSupport vs = provider.getComputeServices().getVolumeSupport();
+
+        if( vs != null ) {
+            Volume volume = vs.getVolume(volumeId);
+
+            if( volume != null ) {
+                long timeout = System.currentTimeMillis() + (CalendarWrapper.MINUTE*20L);
+
+                while( timeout > System.currentTimeMillis() ) {
+                    try { Thread.sleep(15000L); }
+                    catch( InterruptedException ignore ) { }
+                    try { volume = vs.getVolume(volumeId); }
+                    catch( Throwable ignore ) { }
+                    if( volume == null || volume.getCurrentState().equals(VolumeState.AVAILABLE) || volume.getCurrentState().equals(VolumeState.DELETED) ) {
+                        break;
+                    }
+                }
+            }
+            if( volume != null && volume.getProviderVirtualMachineId() == null && support.identifyAttachmentRequirement().equals(Requirement.REQUIRED) ) {
+                String vmId = getTestVmId(DaseinTestManager.STATEFUL, VmState.RUNNING, true, volume.getProviderDataCenterId());
+
+                if( vmId != null ) {
+                    @SuppressWarnings("ConstantConditions") VirtualMachine vm = provider.getComputeServices().getVirtualMachineSupport().getVirtualMachine(vmId);
+
+                    if( vm != null ) {
+                        for( String deviceId : vs.listPossibleDeviceIds(vm.getPlatform()) ) {
+                            try {
+                                vs.attach(volumeId, vmId, deviceId);
+                                break;
+                            }
+                            catch( Throwable ignore ) {
+                                // ignore
+                            }
+                        }
+                    }
+                }
             }
         }
         options = SnapshotCreateOptions.getInstanceForCreate(volumeId, namePrefix + (System.currentTimeMillis()%10000), "Dasein Snapshot Test " + label);
